+++ CHANGES.md
@@ -63,6 +63,7 @@ Changes in 13.0
 - Add link to rosariosis.org/modules & rosariosis.org/plugins in Modules.inc.php & Plugins.inc.php
 - Security fix check sibling belongs to the authenticated parent in Registration.php & Registration.fnc.php, thanks to Kai Aizen
 - Do not save violations triggered by "Check your Network" AJAX error in SaveReport.php
+- HTML add "Calendars" fieldset in Configuration.php
 
 Changes in 12.9.3
 -----------------